Add support for using tags.
Tags have to be specified by prefixing the name of the corresponding row
with '_tag'.
To make the usage easier tags are first matched case-sensitive and if
nothing is found case-insensitive.
Since tags have to be created by an administrator not having
corresponding tags is seen as a failure and will lead to errors.